// ferrum-admin/client/src/endpoints/connection_client.ts
// This file was automatically generated by generate_client.ts. Want to regenerate it? Run `npm run create:client`


export class ConnectionClient {
    public static async addServer(body: {
            name: string;
            serverIP: string;
            serverPort: number;
        }): Promise<{
            success: boolean;
            error?: string;
        }> {
        const request = await fetch('/api/connect/addServer', {
            method: 'POST',
            body: JSON.stringify(body),
            headers: {
                'Content-Type': 'application/json'
            }
        });

        if (!request.ok) {
            throw new Error('Request failed');
        }
        return request.json();
    }
    public static async removeServer(body: {
            serverIP: string;
            serverPort: number;
        }): Promise<{
            success: boolean;
        }> {
        const request = await fetch('/api/connect/removeServer', {
            method: 'POST',
            body: JSON.stringify(body),
            headers: {
                'Content-Type': 'application/json'
            }
        });

        if (!request.ok) {
            throw new Error('Request failed');
        }
        return request.json();
    }
    public static async listServers(): Promise<{ servers: { serverIP: string; serverPort: number; name: string }[] }> {
        const request = await fetch('/api/connect/listServers');
        if (!request.ok) {
            throw new Error('Request failed');
        }

        return request.json();
    }
}